Add NimMain to binding wrapper api
with importc

when cpp backend generates library,
it mangles NimMain for static lib,
but doesn't mangle for dynamic lib.
So NimMain inside extern C works for dynamic lib,
but for static lib, it has to be outside of extern C.

Declaring it as exportc in binding api stop the mangling,
so it works in extern C for static lib too.

Signed-off-by: atunu <atunutemp1@gmail.com>